feat: Add debouncing to avoid false stops
I've been seeing the service get stopped even when I know my machine is not
in use (maybe the cat bumped the mouse?). Add some hacky debouncing to
avoid stopping the service if we're actually still idle.

This is implemented by sleeping for 5 seconds and seeing if we were idle
for that entire time. If not, we consider this transition from idle to
non-idle a fluke and (hackily) ignore it.